requirements.txt
PyGithub
trufflehog
